% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Predict.matrix.pcre.random.effect}
\alias{Predict.matrix.pcre.random.effect}
\title{mgcv-style constructor for prediction of PC-basis functional random effects}
\usage{
\method{Predict.matrix}{pcre.random.effect}(object, data)
}
\arguments{
\item{object}{a smooth specification object, see \code{\link[mgcv]{smooth.construct}}}

\item{data}{see \code{\link[mgcv]{smooth.construct}}}
}
\value{
design matrix for PC-based functional random effects
}
\description{
mgcv-style constructor for prediction of PC-basis functional random effects
}
\author{
Fabian Scheipl;  adapted from 'Predict.matrix.random.effect' by S.N. Wood.
}

