\name{refund-internal}
\alias{decomp}
\alias{decomp2d}
\alias{decomp3d}
\alias{reconstr}
\alias{reconstr2d}
\alias{reconstr3d}
\alias{fpcr.setup}
\alias{first.last_test}
\alias{getSpandDist}
\alias{getNPC.DonohoGavish}
\alias{imwd_test}
\alias{osplinepen2d}
\alias{Omegas}
\alias{lw.test}
\alias{safeDeparse}
\alias{getShrtlbls}
\alias{list2df}
\alias{preprocess.pfr}
\alias{postprocess.pfr}
\alias{parse.predict.pfr}
\alias{Predict.matrix.pss.smooth}
\alias{checkError}
\alias{getRsq}
\alias{decorrelate}
\alias{waveletGetResult}
\alias{waveletSetup}
\alias{pspline.setting}
\alias{waveletGetCV}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal functions for the refund package
}
\description{
These functions are ordinarily not to be called by the user, but if you contact the package authors with any questions about them, we'll do our best to clarify matters.}


