\name{refund-package}
\Rdversion{1.1}
\alias{refund-package}
\alias{refund}
\docType{package}
\title{
Regression with Functional Data
}
\description{
Methods for regression with functional data. Various approaches to regression with scalar responses and functional predictors are implemented by \code{\link{pfr}}, \code{\link{peer}}, \code{\link{lpeer}}, \code{\link{fpcr}}, \code{\link{wcr}} and \code{\link{wnet}}. For regression with functional responses, see \code{\link{pffr}}, \code{\link{fosr}}, and \code{\link{fosr2s}}.}
\details{For a complete list of functions type \code{library(help=refund)}.
}

\author{
\emph{Authors:}
Ciprian Crainiceanu \email{ccrainic@jhsph.edu}, Philip Reiss \email{phil.reiss@nyumc.org}, Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}, Lei Huang \email{huangracer@gmail.com}, Lan Huo \email{lan.huo@nyumc.org}, Fabian Scheipl \email{fabian.scheipl@stat.uni-muenchen.de}, Bruce Swihart \email{bruce.swihart@gmail.com}, Luo Xiao \email{lxiao@jhsph.edu}.


\emph{Contributors:}
Sonja Greven, Jarek Harezlak, Madan Gopal Kundu, Matt McLean, Yihong Zhao

\emph{Maintainer:} Lei Huang \email{huangracer@gmail.com}
}
\keyword{ package }

